backend/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
yt-dlp==2025.10.14
python-multipart==0.0.20
stripe==15.0.0
sqlalchemy==2.0.48
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
